style: rename and split out models
models now (mostly) have their own files, rather than being grouped
this doesn't change the code, just makes them easier to find and compare
