+++ NeverBounceSDK/Models/Account/CreditsInfo.cs
@@ -1,0 +1,12 @@
+﻿namespace NeverBounce.Models;
+
+public class CreditsInfo
+{
+    public int PaidCreditsUsed { get; set; }
+
+    public int FreeCreditsUsed { get; set; }
+
+    public int PaidCreditsRemaining { get; set; }
+
+    public int FreeCreditsRemaining { get; set; }
+}
