+++ NeverBounceSDK/Models/Job/JobCreateResponseModel.cs
@@ -1,0 +1,6 @@
+﻿namespace NeverBounce.Models;
+
+public class JobCreateResponseModel : ResponseModel
+{
+    public int JobID { get; set; }
+}
