+++ NeverBounceSDK/Models/Job/JobParseRequestModel.cs
@@ -1,0 +1,9 @@
+﻿namespace NeverBounce.Models;
+
+public class JobParseRequestModel : JobRequestModel
+{
+    public JobParseRequestModel(int jobID) : base(jobID) { }
+
+    /// <summary>Should the job start processing immediately after it's parsed? (default: false)</summary>
+    public bool AutoStart { get; set; } = false;
+}
