+++ NeverBounceSDK/Models/ResponseModel.cs
@@ -1,46 +1,17 @@
 ﻿namespace NeverBounce.Models;
 
-public enum ResponseStatus {
-    None = 0,
-    Success,
-    AuthFailure,
-    TempUnavail,
-    ThrottleTriggered,
-    BadReferrer,
-}
-
+/// <summary>Generic response properties</summary>
 public class ResponseModel
 {
-    public ResponseStatus Status { get; set; } = ResponseStatus.None;
+    /// <summary>All 2xx level responses will contain a status property.
+    /// This property will indicate whether the requested operation was successfully completed or if an error was encountered.
+    /// 
+    /// <para>When an error does occur a message property will be included with a detailed message about why it failed.
+    /// These error messages will be returned with a 200 level status code.</para></summary>
+    public ResponseStatus Status { get; set; } = ResponseStatus.Success;
 
+    /// <summary>When an error does occur a message property will be included with a detailed message about why it faied.</summary>
     public string? Message { get; set; } = null;
 
     public int ExecutionTime { get; set; }
-}
-
-public class RawResponseModel
-{
-    public object Json { get; set; }
-    public string Plaintext { get; set; }
-}
-
-public class CreditsInfo
-{
-    public int PaidCreditsUsed { get; set; }
-    public int FreeCreditsUsed { get; set; }
-    public int PaidCreditsRemaining { get; set; }
-    public int FreeCreditsRemaining { get; set; }
-}
-
-public class AddressInfo
-{
-    public string OriginalEmail { get; set; }
-    public string NormalizedEmail { get; set; }
-    public string Addr { get; set; }
-    public string Alias { get; set; }
-    public string Host { get; set; }
-    public string Fqdn { get; set; }
-    public string Domain { get; set; }
-    public string Subdomain { get; set; }
-    public string Tld { get; set; }
 }
